Solve a error

+++ docker/entrypoint.sh
@@ -18,12 +18,14 @@ echo "Copying serverfiles overrides..."
 rsync --verbose --archive $home/serverfiles_override/ $gmodroot/
 
 # Any additional files
-cp $home/_gluatest_artifacts/_gluatest_artifacts/* $gmodroot/
+if [ -d $home/_gluatest_artifacts/_gluatest_artifacts/ ]; then
+    cp $home/_gluatest_artifacts/_gluatest_artifacts/* $gmodroot/
 
-for file in $gmodroot/*.tar.gz; do \
-    tar --extract --verbose --ungzip --file="$file" --directory="$gmodroot" && \
-    rm --force --verbose "$file"; \
-done
+    for file in $gmodroot/*.tar.gz; do \
+        tar --extract --verbose --ungzip --file="$file" --directory="$gmodroot" && \
+        rm --force --verbose "$file"; \
+    done
+fi
 
 if [ -f "$gmodroot/custom_requirements.txt" ]; then
     echo "Appending custom requirements"
